+++ src/main/java/eu/solven/holymolap/aggregate/RawPointToAggregates.java
@@ -1,10 +1,10 @@
 package eu.solven.holymolap.aggregate;
 
 public class RawPointToAggregates<T> {
-	public final int[] valueIndexes;
+	public final long[] valueIndexes;
 	public final T keyToAggregates;
 
-	public RawPointToAggregates(T keyToAggregates, int[] valueIndexes) {
+	public RawPointToAggregates(T keyToAggregates, long[] valueIndexes) {
 		this.keyToAggregates = keyToAggregates;
 		this.valueIndexes = valueIndexes;
 	}
